feat: add work experinece in about page
